+++ fritzconnection/core/fritzconnection.py
@@ -10,6 +10,7 @@ changelog v2.0:
 """
 
 import os
+import string
 
 import requests
 
@@ -17,6 +18,7 @@ from pathlib import Path
 from requests.auth import HTTPDigestAuth
 from typing import Any
 
+from fritzconnection import __version__
 from fritzconnection.core.exceptions import FritzServiceError
 from fritzconnection.core.fritzdescription import FritzDescription
 from fritzconnection.core.fritzhttp import FritzHttp
@@ -27,9 +29,11 @@ FRITZ_IP_ADDRESS = "169.254.1.1"
 FRITZ_TCP_PORT = 49000
 FRITZ_TLS_PORT = 49443
 FRITZ_USERNAME = "dslf-config"  # for Fritz!OS < 7.24
+FRITZ_USERNAME_REQUIRED_VERSION = 7.24
 
 FRITZ_ENV_IPADDRESS = "FRITZ_IPADDRESS"
 FRITZ_ENV_PORT = "FRITZ_PORT"
+FRITZ_ENV_TLS_PORT = "FRITZ_TLS_PORT"
 FRITZ_ENV_USERNAME = "FRITZ_USERNAME"
 FRITZ_ENV_PASSWORD = "FRITZ_PASSWORD"
 FRITZ_ENV_CACHEDIRECTORY = "FRITZ_CACHEDIRECTORY"
@@ -137,6 +141,10 @@ class FritzConnection:
     data (i.e. phone numbers) in debug outputs. Default is `False`.
 
     .. versionadded:: 1.15
+    
+    `use_cache` now defaults to `True`. `cache_format` has been removed.
+    
+    .. versionadded:: 2.0
 
     """
 
@@ -158,7 +166,7 @@ class FritzConnection:
         self.user = get_argument(user, FRITZ_ENV_USERNAME, FRITZ_USERNAME)
         self.password = get_argument(password, FRITZ_ENV_PASSWORD, "")
         if use_tls:
-            port = get_argument(port, FRITZ_ENV_PORT, FRITZ_TLS_PORT)
+            port = get_argument(port, FRITZ_ENV_TLS_PORT, FRITZ_TLS_PORT)
             protocol = "https://"
         else:
             port = get_argument(port, FRITZ_ENV_PORT, FRITZ_TCP_PORT)
@@ -204,6 +212,48 @@ class FritzConnection:
         #self._reset_user(user, password)
         # provide the http-interface
         self.http_interface = FritzHttp(self)
+        
+    def __str__(self):
+        return (
+            f"{self.__class__.__name__} [version {__version__}]\n"
+            f"Device: {self.device_name}\n"
+            f"System: {self.system_version}"
+        )
+
+    def _reset_user(self, user, password):
+        """
+        For Fritz!OS >= 7.24: if a password is given and the username is
+        the historic FRITZ_USERNAME, then check for the last logged-in
+        username and use this username for the soaper. Also recreate the
+        session used by the soaper and the device_manager.
+
+        This may not guarantee a valid user/password combination, but is
+        the way AVM recommends setting the required username in case a
+        username is not provided.
+        """
+        try:
+            sys_version = float(self.system_version)
+        except (ValueError, TypeError):
+            # version not available: don't do anything
+            return
+        if (sys_version >= FRITZ_USERNAME_REQUIRED_VERSION
+            and user == FRITZ_USERNAME
+            and password
+        ):
+            last_user = None
+            response = self.call_action(
+                'LANConfigSecurity1', 'X_AVM-DE_GetUserList'
+            )
+            root = ElementTree.fromstring(response['NewX_AVM-DE_UserList'])
+            for node in root:
+                if node.tag == 'Username' and node.attrib['last_user'] == '1':
+                    last_user = node.text
+                    break
+            if last_user is not None:
+                self.session.auth = HTTPDigestAuth(last_user, password)
+                self.soaper.user = last_user
+                self.soaper.session = self.session
+                self.description.session = self.session
 
     @property
     def device_name(self) -> str:
@@ -211,6 +261,16 @@ class FritzConnection:
         Returns the name of the device.
         """
         return self.description.device_name
+        
+    @property
+    def modelname(self):
+        """
+        Returns the name of the device.
+        
+        .. version-deprecated:: 2.0
+           Use :py:func:`device_name` instead.
+        """
+        return self.device_name
 
     @property
     def system_version(self) -> str:
@@ -236,12 +296,44 @@ class FritzConnection:
         will get combined. A keyword argument with the same name as an
         argument in the `arguments` dictionary will overwrite the value
         in the dictionary.
+        
+        The values in the `arguments` dictionary can be of type *str*,
+        *int* or *bool*. (Note: *bool* is provided since 1.3. In former
+        versions booleans must be provided as numeric values: 1, 0).
+
+        Invalid service names will raise a ServiceError and invalid
+        action names will raise an ActionError.
+
+        Legathy-feature: If the service_name does not end with a
+        number-character (like "1"), a "1" gets added by default. If the
+        service_name ends with a colon and a number, the colon gets
+        removed. So i.e. "WLANConfiguration" expands to
+        "WLANConfiguration1" and "WLANConfiguration:2" converts to
+        "WLANConfiguration2". Newer code should avoid this calling style.
+        
+        The method returns a dictionary with argument-names as keys and
+        the corresponding information as values. Numeric and boolean
+        values are converted from strings to Python datatypes.
         """
         arguments = arguments if arguments else {}
         arguments.update(kwargs)
+        service_name = self._get_normalized_service_name(service_name)
         try:
             service = self.description.services[service_name]
         except KeyError:
             raise FritzServiceError(f'unknown service: "{service_name}"')
         return self.soaper.execute(service, action_name, arguments)
-
+    
+    @staticmethod
+    def _get_normalized_service_name(name):
+        """
+        Returns the normalized service name, i.e. `WLANConfiguration` and
+        `WLANConfiguration:1` will get converted to `WLANConfiguration1`.
+        """
+        if ":" in name:
+            name, number = name.split(":", 1)
+            name = f"{name}{number}"
+        elif name[-1] not in string.digits:
+            name = f"{name}1"
+        return name
+