+++ noxfile.py
@@ -5,15 +5,32 @@ PYTHON_TEST_VERSIONS = ("3.10", "3.11", "3.12", "3.13", "3.14")
 PYTHON_DEVELOPMENT_VERSION = "3.13"
 
 
-@nox.session(python=PYTHON_TEST_VERSIONS)
+@nox.session(python=PYTHON_DEVELOPMENT_VERSION)
 def test(session):
+    session.install("-e", ".")
+    session.install("requests", "pytest")
+    # for *session.posargs:
+    # provide the files to test after a double dash, like
+    # nox -s test -- path/to/testfile
+    session.run("pytest", "-m", "not routertest", *session.posargs)
+
+
+@nox.session(python=PYTHON_TEST_VERSIONS)
+def test_versions(session):
     session.install("-e", ".")
     session.install("requests", "pytest")
     session.run("pytest", "-m", "not routertest", *session.posargs)
 
 
-@nox.session(python=PYTHON_TEST_VERSIONS)
+@nox.session(python=PYTHON_DEVELOPMENT_VERSION)
 def test_router(session):
+    session.install("-e", ".")
+    session.install("requests", "pytest")
+    session.run("pytest", "-m", "routertest")
+
+
+@nox.session(python=PYTHON_TEST_VERSIONS)
+def test_router_versions(session):
     session.install("-e", ".")
     session.install("requests", "pytest")
     session.run("pytest", "-m", "routertest")